App now evaluates ruby expressions and gives answer back
I used http://citizen428.net/blog/2010/05/12/a-poor-mans-repl-in-ruby to create a 'poor man's repl.'  This repl uses ruby's evaluate expression to execute ruby code and send back an answer.

This is totally insecure.  It does not sanitize data at all, and the entire point is to be able to recognize ruby code and run it.  This means that if you text `rm -rf .` to my server, there is a very good chance it will die a horrible death.  So please don't be mean.